requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gunicorn>=21.2
psycopg2-binary>=2.9
